Refactor: split arrow from ops_arrow
