+++ weave/language_features/tagging/make_tag_getter_op.py
@@ -26,7 +26,7 @@ def make_tag_getter_op(
     Returns:
         The op.
     """
-    from ...ops_arrow.list_ import ArrowWeaveList, ArrowWeaveListType
+    from ...arrow.list_ import ArrowWeaveList, ArrowWeaveListType
 
     # Uncomment below once we allow searching by type alone.
     # if tag_key is None and tag_type is None:
