+++ weave/language_features/tagging/process_opdef_resolve_fn.py
@@ -14,6 +14,8 @@ from .tagged_value_type import TaggedValueType
 from ... import box
 from ... import weave_types as types
 from . import tag_store
+from ...arrow.arrow_tags import awl_add_arrow_tags
+from ...arrow.list_ import ArrowWeaveList
 from .opdef_util import (
     get_first_arg,
     should_flow_tags,
@@ -64,9 +66,6 @@ def propagate_arrow_tags(
     args: list[typing.Any],
     kwargs: dict[str, typing.Any],
 ) -> typing.Any:
-    from ...ops_arrow.arrow_tags import awl_add_arrow_tags
-    from ...ops_arrow import ArrowWeaveList
-
     tag_type: typing.Optional[types.Type]
 
     _, first_arg_val = get_first_arg(op_def, args, kwargs)
