+++ weave/ops_arrow/__init__.py
@@ -2,9 +2,9 @@ from .. import context_state as _context_state
 
 _loading_builtins_token = _context_state.set_loading_built_ins()
 
-from .arrow import *
-from .list_ import *
-from .convert import *
+from ..arrow.arrow import *
+from ..arrow.list_ import *
+from ..arrow.convert import *
 from .vectorize import vectorize
 from . import ops
 from .list_range import range
