+++ weave/ops_arrow/number.py
@@ -6,7 +6,7 @@ import datetime
 from ..decorator_arrow_op import arrow_op
 from .. import weave_types as types
 from ..api import op
-from .list_ import ArrowWeaveList, ArrowWeaveListType
+from ..arrow.list_ import ArrowWeaveList, ArrowWeaveListType
 from .. import timestamp as weave_timestamp
 from . import util
