+++ weave/ops_domain/run_history/run_history_v2_parquet_media.py
@@ -11,7 +11,9 @@ from ... import artifact_mem
 from .. import wb_util
 from ...ops_domain import wbmedia
 from ...ops_arrow.list_ops import concat
-from ...ops_arrow import ArrowWeaveList, ArrowWeaveListType, convert
+from ...arrow.list_ import ArrowWeaveList
+from ...arrow.list_ import ArrowWeaveListType
+from ...arrow import convert
 from ... import engine_trace
 from ... import gql_json_cache
