--- weave/test_mol.pdb
@@ -1,1 +1,0 @@
-00000
